Refactored swap/scratch to state
