+++ src/state/mod.rs
@@ -1,10 +1,37 @@
+#[cfg(feature = "ram-state")]
+pub mod ram;
+
+#[cfg(feature = "scratch-state")]
+pub mod scratch;
+
+use embedded_storage::Storage;
+
+/// Error occured during memory access
+#[cfg_attr(feature = "use-defmt", derive(Format))]
+#[derive(Debug)]
+pub enum MemoryError {
+    BankSizeNotEqual,
+    BankSizeZero,
+    ReadFailure,
+    WriteFailure,
+}
+
+/// Abstraction for the exchange operation of the current state.
+pub trait Exchange {
+    fn exchange<STORAGE: Storage, STATE: State, const INTERNAL_PAGE_SIZE: usize>(
+        &mut self,
+        internal_memory: &mut STORAGE,
+        state: &mut STATE,
+        progress: ExchangeProgress,
+    ) -> Result<(), MemoryError>;
+}
+
 use crate::hardware::Bank;
-use crate::log;
 
 use crc::{Crc, CRC_32_CKSUM};
 #[cfg(feature = "defmt")]
 use defmt::Format;
-#[cfg(feature = "ram-state")]
+#[cfg(any(feature = "ram-state", feature = "scratch-state"))]
 use desse::{Desse, DesseSized};
 #[cfg(feature = "serde")]
 use serde::{Deserialize, Serialize};
@@ -14,7 +41,10 @@ use serde::{Deserialize, Serialize};
 // exchanged via software updates easily
 #[cfg_attr(feature = "use-defmt", derive(Format))]
 #[cfg_attr(feature = "derive", derive(Serialize, Deserialize))]
-#[cfg_attr(feature = "ram-state", derive(Desse, DesseSized))]
+#[cfg_attr(
+    any(feature = "ram-state", feature = "scratch-state"),
+    derive(Desse, DesseSized)
+)]
 #[derive(Debug, PartialEq, Eq)]
 pub enum Update {
     // No update requested, just jump to the application
@@ -33,7 +63,10 @@ pub enum Update {
 
 #[cfg_attr(feature = "use-defmt", derive(Format))]
 #[cfg_attr(feature = "derive", derive(Serialize, Deserialize))]
-#[cfg_attr(feature = "ram-state", derive(Desse, DesseSized))]
+#[cfg_attr(
+    any(feature = "ram-state", feature = "scratch-state"),
+    derive(Desse, DesseSized)
+)]
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 /// Errors that can occur during update
 pub enum UpdateError {
@@ -50,7 +83,10 @@ pub enum UpdateError {
 /// Upcoming operation of the exchange process for a given page index.
 #[cfg_attr(feature = "use-defmt", derive(Format))]
 #[cfg_attr(feature = "derive", derive(Serialize, Deserialize))]
-#[cfg_attr(feature = "ram-state", derive(Desse, DesseSized))]
+#[cfg_attr(
+    any(feature = "ram-state", feature = "scratch-state"),
+    derive(Desse, DesseSized)
+)]
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub enum ExchangeStep {
     /// Copy page A to the scratch page.
@@ -64,7 +100,10 @@ pub enum ExchangeStep {
 /// Store the progress of the current exchange operation
 #[cfg_attr(feature = "use-defmt", derive(Format))]
 #[cfg_attr(feature = "derive", derive(Serialize, Deserialize))]
-#[cfg_attr(feature = "ram-state", derive(Desse, DesseSized))]
+#[cfg_attr(
+    any(feature = "ram-state", feature = "scratch-state"),
+    derive(Desse, DesseSized)
+)]
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub struct ExchangeProgress {
     /// Bank the update is coming from
@@ -82,7 +121,10 @@ pub struct ExchangeProgress {
 /// Struct used to store the state of the bootloader situation in NVM
 #[cfg_attr(feature = "use-defmt", derive(Format))]
 #[cfg_attr(feature = "derive", derive(Serialize, Deserialize))]
-#[cfg_attr(feature = "ram-state", derive(Desse, DesseSized))]
+#[cfg_attr(
+    any(feature = "ram-state", feature = "scratch-state"),
+    derive(Desse, DesseSized)
+)]
 #[derive(Debug)]
 pub struct MoonbootState {
     /// If set Request, an Update is requested. This will exchange the two images, set the update
@@ -111,64 +153,4 @@ const CRC: Crc<StateCrcType> = Crc::<StateCrcType>::new(&CRC_32_CKSUM);
 
 fn checksum(bytes: &[u8]) -> StateCrcType {
     CRC.checksum(bytes)
-}
-
-/// State stored in the RAM
-/// TODO: Move to hardware folder together with state trait?
-#[cfg(feature = "ram-state")]
-pub mod ram {
-    use super::*;
-
-    /// State read and written to RAM. This assumes the device is never powered off / the ram is never
-    /// reset!
-    pub struct RamState;
-
-    extern "C" {
-        static mut _moonboot_state_crc_start: StateCrcType;
-        static mut _moonboot_state_data_start: [u8; STATE_SERIALIZED_MAX_SIZE];
-        // TODO: Move these as normal variables to linker sections via #[link] macro?
-    }
-
-    impl State for RamState {
-        fn read(&mut self) -> MoonbootState {
-            let crc = unsafe { _moonboot_state_crc_start };
-
-            log::info!(
-                "Reading data with len: {}, CRC: {}",
-                STATE_SERIALIZED_MAX_SIZE,
-                crc
-            );
-
-            let checksum = checksum(unsafe { &_moonboot_state_data_start });
-            if crc == checksum {
-                let data = MoonbootState::deserialize_from(unsafe { &_moonboot_state_data_start });
-                log::trace!("CRC Match! {}: {:?}", crc, data);
-                return data;
-            } else {
-                log::trace!("CRC Mismatch! {} vs {}", crc, checksum);
-            }
-
-            MoonbootState {
-                update: Update::None,
-            }
-        }
-
-        fn write(&mut self, data: &MoonbootState) -> Result<(), ()> {
-            log::trace!("Writing data {:?}", data);
-
-            unsafe { _moonboot_state_data_start = data.serialize() };
-            log::trace!("Written data: {:?}", unsafe { &_moonboot_state_data_start });
-
-            unsafe {
-                _moonboot_state_crc_start = checksum(&_moonboot_state_data_start);
-            }
-            log::info!(
-                "Written len: {}, checksum: {}",
-                STATE_SERIALIZED_MAX_SIZE,
-                unsafe { _moonboot_state_crc_start }
-            );
-
-            Ok(())
-        }
-    }
 }
