+++ src/state/scratch.rs
